+++ xclim/indicators/generic/_stats.py
@@ -11,6 +11,7 @@ fit = Indicator(
     title="Distribution parameters fitted over the time dimension.",
     identifier="fit",
     var_name="params",
+    units="",
     standard_name="{dist} parameters",
     long_name="{dist} distribution parameters",
     description="Parameters of the {dist} distribution.",