+++ xclim/indicators/land/_streamflow.py
@@ -7,13 +7,15 @@ from xclim.core.cfchecks import check_valid
 from xclim.core.indicator import Indicator, ResamplingIndicator
 from xclim.core.units import declare_units
 from xclim.indices import base_flow_index, generic, rb_flashiness_index
+from xclim.indices.stats import fit as _fit
 from xclim.indices.stats import frequency_analysis
 
 __all__ = [
     "base_flow_index",
-    "discharge_return_level",
     "rb_flashiness_index",
-    "discharge_stats",
+    "freq_analysis",
+    "stats",
+    "fit",
     "doy_qmax",
     "doy_qmin",
 ]
@@ -38,7 +40,7 @@ base_flow_index = Streamflow(
     compute=base_flow_index,
 )
 
-discharge_return_level = Streamflow(
+freq_analysis = Streamflow(
     title="Return level",
     identifier="freq_analysis",
     var_name="q{window}{mode:r}{indexer}",
@@ -50,6 +52,7 @@ discharge_return_level = Streamflow(
     compute=frequency_analysis,
     missing="skip",
     input={"da": "discharge"},
+    _version_deprecated="0.40",
 )
 
 rb_flashiness_index = Streamflow(
@@ -65,7 +68,7 @@ rb_flashiness_index = Streamflow(
 )
 
 
-discharge_stats = Streamflow(
+stats = Streamflow(
     title="Statistic of the daily flow for a given period.",
     identifier="discharge_stats",
     var_name="q{indexer}{op:r}",
@@ -75,6 +78,24 @@ discharge_stats = Streamflow(
     compute=generic.select_resample_op,
     missing="any",
     input={"da": "discharge"},
+    _version_deprecated="0.40",
+)
+
+
+fit = Indicator(
+    title="Distribution parameters fitted over the time dimension.",
+    identifier="fit",
+    var_name="params",
+    units="",
+    standard_name="{dist} parameters",
+    long_name="{dist} distribution parameters",
+    description="Parameters of the {dist} distribution.",
+    cell_methods="time: fit",
+    src_freq=None,
+    compute=_fit,
+    input={"da": "discharge"},
+    realm="land",
+    _version_deprecated="0.40",
 )