+++ xclim/testing/tests/test_land.py
@@ -6,7 +6,7 @@ import pytest
 import xarray as xr
 
 import xclim.core.utils
-from xclim import generic, land
+from xclim import land
 
 
 def test_base_flow_index(ndq_series):
@@ -23,7 +23,7 @@ def test_rb_flashiness_index(ndq_series):
 
 class Test_FA:
     def test_simple(self, ndq_series):
-        out = land.discharge_return_level(
+        out = land.freq_analysis(
             ndq_series, mode="max", t=[2, 5], dist="gamma", season="DJF"
         )
         assert out.description in [
@@ -35,9 +35,7 @@ class Test_FA:
         np.testing.assert_array_equal(out.isnull(), False)
 
     def test_no_indexer(self, ndq_series):
-        out = land.discharge_return_level(
-            ndq_series, mode="max", t=[2, 5], dist="gamma"
-        )
+        out = land.freq_analysis(ndq_series, mode="max", t=[2, 5], dist="gamma")
         assert out.description in [
             "Streamflow frequency analysis for the maximal annual 1-day flow "
             "estimated using the gamma distribution."
@@ -47,36 +45,34 @@ class Test_FA:
         np.testing.assert_array_equal(out.isnull(), False)
 
     def test_q27(self, ndq_series):
-        out = land.discharge_return_level(
-            ndq_series, mode="max", t=2, dist="gamma", window=7
-        )
+        out = land.freq_analysis(ndq_series, mode="max", t=2, dist="gamma", window=7)
         assert out.shape == (1, 2, 3)
 
     def test_empty(self, ndq_series):
         q = ndq_series.copy()
         q[:, 0, 0] = np.nan
-        out = land.discharge_return_level(
+        out = land.freq_analysis(
             q, mode="max", t=2, dist="genextreme", window=6, freq="YS"
         )
         assert np.isnan(out.values[:, 0, 0]).all()
 
     def test_wrong_variable(self, pr_series):
         with pytest.raises(xclim.core.utils.ValidationError):
-            land.discharge_return_level(
+            land.freq_analysis(
                 pr_series(np.random.rand(100)), mode="max", t=2, dist="gamma"
             )
 
 
 class TestStats:
     def test_simple(self, ndq_series):
-        out = land.discharge_stats(ndq_series, freq="YS", op="min", season="MAM")
+        out = land.stats(ndq_series, freq="YS", op="min", season="MAM")
         assert out.attrs["units"] == "m^3 s-1"
 
     def test_missing(self, ndq_series):
         a = ndq_series
         a = ndq_series.where(~((a.time.dt.dayofyear == 5) * (a.time.dt.year == 1902)))
         assert a.shape == (5000, 2, 3)
-        out = land.discharge_stats(a, op="max", month=1)
+        out = land.stats(a, op="max", month=1)
 
         np.testing.assert_array_equal(out.sel(time="1900").isnull(), False)
         np.testing.assert_array_equal(out.sel(time="1902").isnull(), True)